Hotfix for detection of Pyrite V1 drives
Pyrite drives with V1 feature set would cause the program to
not recognize the feature set and cause an unknown error with
Discovery 1, 2 and rng. This fixes the detection of Pyrite V1, but
future changes will be added to differentiate between Pyrite V1
and V2 drives.

Note: This can be done, because Pyrite V1 and V2 feature sets
have identical schemas.